Make it build on Snow Leopard. Build a universal 32/64 bit binary for release. Switch to already open tab when opening files.

+++ MarkerLineNumberView.m
@@ -42,7 +42,7 @@
 	[super dealloc];
 }
 
-- (void)setRuleThickness:(float)thickness
+- (void)setRuleThickness:(CGFloat)thickness
 {
 	[super setRuleThickness:thickness];
 	
@@ -96,8 +96,8 @@
 - (void)mouseDown:(NSEvent *)theEvent
 {
 	NSPoint					location;
-	unsigned				line;
-	
+	NSUInteger				line;
+
 	location = [self convertPoint:[theEvent locationInWindow] fromView:nil];
 	line = [self lineNumberForLocation:location.y];
