+++ ViRegexp.h
@@ -13,8 +13,8 @@
 + (ViRegexpMatch *)regexpMatchWithRegion:(OnigRegion *)aRegion startLocation:(NSUInteger)aLocation;
 - (ViRegexpMatch *)initWithRegion:(OnigRegion *)aRegion startLocation:(NSUInteger)aLocation;
 - (NSRange)rangeOfMatchedString;
-- (NSRange)rangeOfSubstringAtIndex:(unsigned)index;
-- (unsigned)count;
+- (NSRange)rangeOfSubstringAtIndex:(NSUInteger)index;
+- (NSUInteger)count;
 
 @end
